delete views from central_erros
