+++ central_erros/urls.py
@@ -16,10 +16,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 
-from central_erros.views import list_errors
-
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('events/', include(('events.urls', 'events'), namespace='events')),
-    path('', list_errors),
 ]
